refactor(back): move BotPos
